+++ package.json
@@ -16,6 +16,7 @@
     "body-parser": "~1.17.1",
     "codecov": "^2.3.0",
     "cookie-parser": "~1.4.3",
+    "cookie-session": "^1.3.1",
     "debug": "~2.6.3",
     "env2": "^2.2.0",
     "express": "~4.15.2",
